New module linkstatus.
